refactor: 도메인 별 패키지 구조로 변경
